Move shared Unix drivers to oshi.driver.common.unix.*
Sweeps the older Unix ports to match the AIX convention established in
#3350. The shared (oshi-common) drivers for BSD, FreeBSD, and Solaris
were sitting at the top-level `oshi.driver.unix.<os>` namespace, which
worked only because the JNA-only drivers in oshi-core happened to live
in disjoint sub-packages (e.g. oshi-core's `oshi.driver.unix.solaris`
holds `PsInfo`/`Who`, oshi-common's `oshi.driver.unix.solaris.disk`
holds `Iostat`/`Lshal`/`Prtvtoc` — different Java packages, no split).
The new convention puts every oshi-common driver under
`oshi.driver.common.unix.<os>[.sub]`, mirroring `oshi.driver.common.mac`
and `oshi.driver.common.windows.*` and avoiding the
top-level-namespace-shared-across-modules fragility.

Moved (oshi-common):
* oshi.driver.unix.bsd.Systat -> oshi.driver.common.unix.bsd.Systat
* oshi.driver.unix.bsd.disk.Disklabel
  -> oshi.driver.common.unix.bsd.disk.Disklabel
* oshi.driver.unix.freebsd.disk.{GeomDiskList,GeomPartList,Mount}
  -> oshi.driver.common.unix.freebsd.disk.*
* oshi.driver.unix.solaris.disk.{Iostat,Lshal,Prtvtoc}
  -> oshi.driver.common.unix.solaris.disk.*
* Tests follow their drivers under
  oshi-common/src/test/java/oshi/driver/common/unix/...

Updated:
* Consumer imports in 8 hardware classes across oshi-common, oshi-core,
  and oshi-core-ffm (FreeBsd/OpenBsd/NetBsd/Solaris HW disk/sensors/
  power source FFM + JNA peers).
* oshi-common module-info.java exports: the four old `oshi.driver.unix.*`
  entries replaced by their `oshi.driver.common.unix.*` equivalents,
  sorted alphabetically.
* oshi-common module-info.test add-opens: matching package renames so
  JUnit Platform can still introspect each test class.
* oshi-core and oshi-core-ffm bnd Export-Package globs: the moved
  packages are now covered by the existing `!oshi.driver.common.*`
  exclude, so the explicit
  `!oshi.driver.unix.{bsd,freebsd.disk,solaris.disk}.*` overrides drop
  out.

Pure refactor — no public API surface changes (everything in the
moved packages was already shared between oshi-common consumers; the
exported package names change but the class names don't). Full
`mvn clean install` green through oshi-dist; bnd manifests show the
new export paths with no leaks of the old paths.
