+++ oshi-common/src/main/java/module-info.java
@@ -22,14 +22,14 @@ module com.github.oshi.common {
     exports oshi.annotation.concurrent;
     exports oshi.driver.common.mac;
     exports oshi.driver.common.unix.aix;
+    exports oshi.driver.common.unix.bsd;
+    exports oshi.driver.common.unix.bsd.disk;
+    exports oshi.driver.common.unix.freebsd.disk;
+    exports oshi.driver.common.unix.solaris.disk;
     exports oshi.driver.common.windows.perfmon;
     exports oshi.driver.common.windows.registry;
     exports oshi.driver.common.windows.gpu;
     exports oshi.driver.common.windows.wmi;
-    exports oshi.driver.unix.bsd;
-    exports oshi.driver.unix.bsd.disk;
-    exports oshi.driver.unix.freebsd.disk;
-    exports oshi.driver.unix.solaris.disk;
     exports oshi.hardware;
     exports oshi.hardware.common;
     exports oshi.hardware.common.platform.linux;
