+++ oshi-common/src/main/java/oshi/driver/common/unix/solaris/disk/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Provides functions to query Solaris disk information.
  */
-package oshi.driver.unix.solaris.disk;
+package oshi.driver.common.unix.solaris.disk;
