+++ oshi-common/src/test/java/module-info.test
@@ -47,13 +47,13 @@
 --add-opens
   com.github.oshi.common/oshi.driver.common.windows.gpu=org.junit.platform.commons
 --add-opens
-  com.github.oshi.common/oshi.driver.unix.bsd=org.junit.platform.commons
+  com.github.oshi.common/oshi.driver.common.unix.bsd=org.junit.platform.commons
 --add-opens
-  com.github.oshi.common/oshi.driver.unix.bsd.disk=org.junit.platform.commons
+  com.github.oshi.common/oshi.driver.common.unix.bsd.disk=org.junit.platform.commons
 --add-opens
-  com.github.oshi.common/oshi.driver.unix.freebsd.disk=org.junit.platform.commons
+  com.github.oshi.common/oshi.driver.common.unix.freebsd.disk=org.junit.platform.commons
 --add-opens
-  com.github.oshi.common/oshi.driver.unix.solaris.disk=org.junit.platform.commons
+  com.github.oshi.common/oshi.driver.common.unix.solaris.disk=org.junit.platform.commons
 --add-opens
   com.github.oshi.common/oshi.hardware.common.platform.unix.freebsd=org.junit.platform.commons
 --add-opens
