+++ oshi-common/src/test/java/oshi/driver/common/unix/solaris/disk/DiskDriversTest.java
@@ -2,7 +2,7 @@
  * Copyright 2022-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.driver.unix.solaris.disk;
+package oshi.driver.common.unix.solaris.disk;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.anEmptyMap;
